Activys
Add new activy ,background,Cardview on recicle view.
