+++ app/src/main/java/com/example/macbook/informatec/MainActivity.java
@@ -22,10 +22,6 @@ import java.util.ArrayList;
 
 public class MainActivity extends AppCompatActivity {
 
-    RecyclerView recyclerView;
-    DatabaseReference reference;
-    ArrayList<Events> list;
-    MyAdapter adapter;
     Button btn;
     Button btn2;
 
@@ -50,35 +46,6 @@ public class MainActivity extends AppCompatActivity {
                 startActivity(new Intent(MainActivity.this,MainActivityEventos.class));
             }
         });
-
-        recyclerView =findViewById(R.id.myRecyclerView);
-        recyclerView.setLayoutManager(new LinearLayoutManager(this));
-        list = new ArrayList<Events>();
-
-        reference = FirebaseDatabase.getInstance().getReference().child("Eventos");
-
-        reference.addValueEventListener(new ValueEventListener() {
-            @Override
-            public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
-
-                for (DataSnapshot dataSnapshot1: dataSnapshot.getChildren())
-                {
-
-                    Events p= dataSnapshot1.getValue(Events.class);
-                    list.add(p);
-
-                }
-
-                adapter = new MyAdapter(MainActivity.this,list);
-                recyclerView.setAdapter(adapter);
-            }
-
-            @Override
-            public void onCancelled(@NonNull DatabaseError databaseError) {
-                Toast.makeText(MainActivity.this,"Salio algo mal",Toast.LENGTH_LONG).show();
-            }
-        });
-
