+++ app/src/main/res/layout/activity_main_concursos.xml
@@ -6,4 +6,12 @@
     android:layout_height="match_parent"
     tools:context=".MainActivityConcursos">
 
+    <android.support.v7.widget.RecyclerView
+        android:id="@+id/myRecyclerView3"
+        android:layout_width="388dp"
+        android:layout_height="511dp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 </android.support.constraint.ConstraintLayout>
