+++ app/src/main/res/layout/content.xml
@@ -12,47 +12,51 @@
         android:layout_weight="1"
         android:orientation="vertical">
 
+        <android.support.v7.widget.CardView xmlns:card_view="http://schemas.android.com/apk/res-auto"
+            android:id="@+id/card_view"
+            android:layout_width="384dp"
+            android:layout_height="118dp"
+            android:layout_gravity="center"
+            card_view:cardCornerRadius="4dp">
+
+        </android.support.v7.widget.CardView>
+
         <TextView
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:gravity="center"
             android:text="@string/name"
+            android:textAlignment="center"
             android:textSize="24sp"
             android:textStyle="bold" />
 
         <TextView
             android:id="@+id/text_name"
-            android:layout_width="match_parent"
-            android:layout_height="35dp"
+            android:layout_width="401dp"
+            android:layout_height="wrap_content"
             android:gravity="center"
-            android:textSize="24sp" />
+            android:textAlignment="center"
+            android:textSize="24sp"
+            android:textStyle="italic" />
 
         <TextView
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:gravity="center"
             android:text="@string/ubication"
+            android:textAlignment="center"
             android:textSize="24sp"
             android:textStyle="bold" />
 
         <TextView
             android:id="@+id/ubication"
             android:layout_width="match_parent"
-            android:layout_height="45dp"
+            android:layout_height="wrap_content"
             android:gravity="center"
-            android:textSize="24sp" />
+            android:textAlignment="center"
+            android:textSize="24sp"
+            android:textStyle="italic" />
 
     </LinearLayout>
 
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_weight="1"
-        android:layout_height="match_parent">
-
-        <ImageView
-            android:id="@+id/imagen"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent" />
-
-    </LinearLayout>
 </LinearLayout>
