+++ app/src/main/res/values/colors.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="colorPrimary">#00008577</color>
-    <color name="colorPrimaryDark">#00574B</color>
-    <color name="colorAccent">#D81B60</color>
+    <color name="colorPrimary">#0018ffff</color>
+    <color name="colorPrimaryDark">#0018ffff</color>
+    <color name="colorAccent">#00d81b60</color>
 </resources>
